+++ tower-http/CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-None.
+- Add `ClassifyResponse::map_failure_class` and `ClassifyEos::map_failure_class`
+  for transforming the failure classification using a function.
 
 ## Breaking changes